refactor: relocate evaluation.py + extract reporting.py from harness.py
- src/pdft_benchmarks/evaluation.py (was evaluation.py): metric
  computation (compute_metrics, aggregate_per_keep_ratio,
  evaluate_basis_shared, evaluate_basis_per_image, evaluate_baseline).
- src/pdft_benchmarks/reporting.py (NEW): dump_metrics_json +
  _julia_float_postprocess pulled out of harness.py. Single import for
  Julia-byte-compat metrics output.

harness.py keeps a re-export of dump_metrics_json for back-compat with
any out-of-tree consumers; the in-repo runners now import from reporting.
