+++ run_quickdraw.py
@@ -38,9 +38,9 @@ from pdft_benchmarks.baselines import (
 )
 from pdft_benchmarks.presets import get_preset
 from pdft_benchmarks.datasets.quickdraw import load_quickdraw
-from evaluation import evaluate_baseline, evaluate_basis_shared
+from pdft_benchmarks.evaluation import evaluate_baseline, evaluate_basis_shared
 from generate_report import main as generate_report_main
-from harness import dump_metrics_json, train_one_basis_batched
+from pdft_benchmarks.reporting import dump_metrics_json, train_one_basis_batched
 
 # ----------------------------------------------------------------------------
 # Constants
